added h1 tags and renamed 2 files
