+++ scripts/switzerland/celebrities/CelebrityDataProvider.js
