+++ scripts/switzerland/main.js
@@ -1,7 +1,7 @@
 import { renderNav } from "../ui/renderNav.js"
 import { SwissCityList } from "./cities/CityList.js"
 import { SwissLandmarkList } from "./landmarks/LandmarkList.js"
-import { SwissCelebrityList } from "./celebrities/CelebritiesList.js"
+import { SwissCelebrityList } from "./celebrities/CelebrityList.js"
 
 renderNav()
 SwissCityList()
